extract-Weather-class
